requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
filelock==3.18.0
fsspec==2025.7.0
hf-xet==1.1.5
huggingface-hub==0.34.3
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
packaging==25.0
pillow==11.3.0
PyYAML==6.0.2
regex==2025.7.31
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.54.1
typing_extensions==4.14.1
urllib3==2.5.0
